feat(schema): implement SchemaValidator for deep merging of multi-file schemas
